提交最新代碼

+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/game/entity/PlayerRegisterInfoVO.java
@@ -47,7 +47,7 @@ public class PlayerRegisterInfoVO implements Serializable {
      * 玩家id
      */
     @Excel(name = "玩家ID", width = 15)
-    private Integer playerId;
+    private Long playerId;
 
     /**
      * 创建日期
